fix widget built twice during warm up frame
