+++ packages/flutter/lib/src/widgets/nested_scroll_view.dart
@@ -498,13 +498,8 @@ class _NestedScrollCoordinator implements ScrollActivityDelegate, ScrollHoldCont
 
   bool get hasScrolledBody {
     for (_NestedScrollPosition position in _innerPositions) {
-      // TODO(chunhtai): Replace null check with assert once
-      // https://github.com/flutter/flutter/issues/31195 is fixed.
-      if (
-        position.minScrollExtent != null &&
-        position.pixels != null &&
-        position.pixels > position.minScrollExtent
-      ) {
+      assert(position.minScrollExtent != null && position.pixels != null);
+      if (position.pixels > position.minScrollExtent) {
         return true;
       }
     }
